orchestrator/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic
structlog
prometheus_client
